backend/requirements.txt
alembic==1.8.1
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
config==0.5.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
flasgger==0.9.5
Flask==2.2.3
Flask-APScheduler==1.12.4
flask-buzz==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
flask-praetorian==1.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
Mako==1.2.4
MarkupSafe==2.1.1
mistune==2.0.4
packaging==23.0
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
psycopg2==2.9.5
py-buzz==2.1.3
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.2.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
redis==4.5.4
requests==2.28.1
sh==2.0.4
six==1.16.0
SQLAlchemy==1.4.44
SQLAlchemy-serializer==1.4.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.10.0
